Add files via upload
Data Science Salary Insights and Excel Dashboard.
The objective of the Data Science Salary Insights and Excel Dashboard is to provide a comprehensive and visually appealing analysis of salary trends in the field of Data Science. The project aims to gather, process, and present relevant salary data to help professionals.